Fix MultiZKConnectionConfig test failure.
Theory - We test MultiZK configuration in two mode, dedicated and MSDS (Zooscalability).
TestMultiZkHelixJavaApi - is focused on MSDS while ConnectionConfig is focused on dedicated ZK.

Test is very unstable wrt "testHelixManager".
If MSDS based tests are run prior to Dedicated ZK, the test fails. we wanted to make sure that Dedicated ZK testcases run prior to MSDS and validate test passes.

Testing Done:
Same test i ran in loop for 10 times:
START TestMultiZkConnectionConfig_testZkRoutingDataSourceConfigs at Tue Apr 11 08:05:59 PDT 2023
END TestMultiZkConnectionConfig_testZkRoutingDataSourceConfigs at Tue Apr 11 08:06:11 PDT 2023
 ======================================================================
 Attempt 10 org.apache.helix.integration.multizk.TestMultiZkConnectionConfig
 ======================================================================
Start zookeeper at localhost:8977 in thread main
Start zookeeper at localhost:8978 in thread main
Start zookeeper at localhost:8979 in thread main
